requirements.txt
adbutils==0.16.2
apkutils2==1.0.0
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.12
cigam==0.0.3
decorator==5.1.1
Deprecated==1.2.13
deprecation==2.1.0
filelock==3.7.0
greenlet==1.1.2
idna==3.3
loguru==0.6.0
logzero==1.7.0
lxml==4.8.0
mimesis==5.4.0
minio==7.1.8
packaging==20.9
Pillow==9.1.0
progress==1.6
psycopg2-binary==2.9.3
py==1.11.0
pydantic==1.9.0
pyelftools==0.28
pyparsing==3.0.9
requests==2.27.1
retry==0.9.2
self==2020.12.3
six==1.16.0
SQLAlchemy==1.4.36
typing_extensions==4.2.0
uiautomator2==2.16.14
urllib3==1.26.9
whichcraft==0.6.1
wrapt==1.14.1
xmltodict==0.13.0
sentry-sdk==0.10.2
httpx==0.24.1
python-dotenv==1.0.0
psutil